services/sentiment-processor/pyproject.toml
[tool.poetry]
name = "sentiment-processor"
version = "3.0.0"
description = "Unified sentiment analysis processor for crypto and gold markets"
authors = ["Sentilyze Team <team@sentilyze.live>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
sentilyze-core = {path = "../../shared", develop = true}
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
google-cloud-aiplatform = "^1.38.0"
google-cloud-pubsub = "^2.18.0"
google-cloud-bigquery = "^3.11.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.0"
redis = "^5.0.0"
httpx = "^0.25.0"
python-json-logger = "^2.0.7"
structlog = "^24.1.0"
google-cloud-storage = "^2.14.0"
google-cloud-firestore = "^2.14.0"

[tool.poetry.group.lite.dependencies]
# Optional dependency for Lite sentiment mode (Google Cloud Natural Language API)
# Install with: `poetry install --with lite`
google-cloud-language = "^2.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^24.0.0"
isort = "^5.13.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
